Refactor shard plan tests
